earth0/requirements.txt
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.22
Flask==2.0.1
Flask-Session==0.3.2
gunicorn==20.1.0
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
kiwisolver==1.1.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.1
meson==0.55.3
mysql-replication==0.21
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pg-chameleon==2.0.14
pickleshare==0.7.5
Pillow==8.2.0
ply==3.11
pproxy==2.3.5
prompt-toolkit==3.0.18
protobuf==3.17.1
psycopg2==2.9.1
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
pydot==1.4.2
Pygments==2.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-telegram-bot==13.7
requests==2.24.0
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==1.1.0